Split 'unlocked' event into 'unlocked_flock' and 'unlocked_lockf'
The 'unlocked' event was using flock for the lock.
However, this feature was meant to be used for e.g. vsftp to wait for uploads to finish, but vsftpd uses lockf/fcntl which is a different type of lock
With the new approach it is now possible to explicitly select for the type of lock to become available.